move forward declaration to simd_fwd.hpp file
