+++ c++/nda/simd.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "simd/simd_fwd.hpp"
 #include "simd/simd.hpp"
 #include "simd/mock_simd.hpp"
 #include "simd/simd_cost.hpp"
