+++ c++/nda/simd/simd_cost.hpp
@@ -1,26 +1,10 @@
 #pragma once
+#include "./simd_fwd.hpp"
 #include "../concepts.hpp"
 
 #include <type_traits>
 
 namespace nda {
-
-  //Forward declaration
-  template <typename ValueType, int Rank, typename Layout, char Algebra, typename ContainerPolicy>
-  class basic_array;
-
-  template <typename ValueType, int Rank, typename Layout, char Algebra, typename AccessorPolicy, typename OwningPolicy>
-  class basic_array_view;
-
-  template <char OP, Array A>
-  struct expr_unary;
-
-  template <char OP, ArrayOrScalar L, ArrayOrScalar R>
-  struct expr;
-
-  template <typename F, Array... A>
-  struct expr_call;
-
   template <typename T>
   concept IsBasicArray = requires {
     []<typename ValueType, int Rank, typename Layout, char Algebra, typename ContainerPolicy>(
